refactor: create DeleteInstanceTrait
