+++ src/HeaderSession.php
@@ -6,14 +6,14 @@ use SessionHandlerInterface;
 use Staffbase\plugins\sdk\Helper\URLHelper;
 use Staffbase\plugins\sdk\Exceptions\SSOAuthenticationException;
 use Staffbase\plugins\sdk\Exceptions\SSOException;
-use Staffbase\plugins\sdk\RemoteCall\DeleteInstanceCallHandlerInterface;
+use Staffbase\plugins\sdk\RemoteCall\DeleteInstanceTrait;
 use Staffbase\plugins\sdk\RemoteCall\RemoteCallInterface;
 use Staffbase\plugins\sdk\SessionHandling\SessionTokenDataTrait;
 use Staffbase\plugins\sdk\SSOData\HeaderSSOData;
 
 class HeaderSession
 {
-    use HeaderSSOData, SessionTokenDataTrait;
+    use HeaderSSOData, SessionTokenDataTrait, DeleteInstanceTrait;
 
     /**
      * Session cookies from staffbase are prefixed with sid_ this should
@@ -21,13 +21,13 @@ class HeaderSession
      */
     private const COOKIE_PREFIX = "sid_";
 
-	/**
-	 * The path for the plugin in the experience studio has the pattern
-	 * `pluginId/instanceId/studio`
-	 */
-	private const EDITOR_STUDIO_PATH = "studio";
+    /**
+     * The path for the plugin in the experience studio has the pattern
+     * `pluginId/instanceId/studio`
+     */
+    private const EDITOR_STUDIO_PATH = "studio";
 
-	/**
+    /**
      * @var String|null $pluginInstanceId the id of the currently used instance.
      */
     private ?string $pluginInstanceId = null;
@@ -48,17 +48,16 @@ class HeaderSession
     private bool $userView;
 
     /**
-     * @var HeaderToken|null token data from the parsed jwt
-     */
-    private ?HeaderToken $sso = null;
-
-
-    /**
      * @throws SSOAuthenticationException
      * @throws SSOException
      */
-    public function __construct(string $pluginId, string $appSecret, ?SessionHandlerInterface $sessionHandler = null, int $leeway = 0, ?RemoteCallInterface $remoteCallHandler = null)
-    {
+    public function __construct(
+        string $pluginId,
+        string $appSecret,
+        ?SessionHandlerInterface $sessionHandler = null,
+        int $leeway = 0,
+        ?RemoteCallInterface $remoteCallHandler = null
+    ) {
         if (!$pluginId) {
             throw new SSOException('Empty plugin ID.');
         }
@@ -68,21 +67,20 @@ class HeaderSession
         }
 
         // we update the SSO info every time we get a token
-        if ($jwt = $this->validateParams($pluginId)) {
-            $this->sso = new HeaderToken($appSecret, $jwt, $leeway);
-        }
+        $sso = ($jwt = $this->validateParams($pluginId)) ? new HeaderToken($appSecret, $jwt, $leeway) : null;
 
         // delete the instance if the special sub is in the token data
-        if ($this->sso && $remoteCallHandler) {
-            $this->deleteInstance($remoteCallHandler);
+        // exits the request
+        if ($sso && $remoteCallHandler && $sso->isDeleteInstanceCall()) {
+            $this->deleteInstance($this->pluginInstanceId, $remoteCallHandler);
         }
 
         $this->storeSessionInfoFromCookie();
 
         $this->openSession($this->session_name, $this->sessionId);
 
-        if ($this->sso !== null) {
-            $this->setClaims($this->sso->getData());
+        if ($sso !== null) {
+            $this->setClaims($sso->getData());
         }
 
         // decide if we are in user view or not
@@ -94,13 +92,13 @@ class HeaderSession
         }
     }
 
-	/**
-	 * Destructor
-	 */
-	public function __destruct()
-	{
-		$this->closeSession();
-	}
+    /**
+     * Destructor
+     */
+    public function __destruct()
+    {
+        $this->closeSession();
+    }
 
     /**
      * Test if userView is enabled.
@@ -158,44 +156,6 @@ class HeaderSession
         $this->sessionId = array_values($cookies)[0];
     }
 
-
-    private function deleteInstance(RemoteCallInterface $remoteCallHandler): void
-    {
-        if (!$this->sso->isDeleteInstanceCall()) {
-            return;
-        }
-
-        $instanceId = $this->pluginInstanceId;
-
-        if ($remoteCallHandler instanceof DeleteInstanceCallHandlerInterface) {
-            $result = $remoteCallHandler->deleteInstance($instanceId);
-        } else {
-            // we will accept unhandled calls with a warning
-            $result = true;
-            error_log("Warning: An instance deletion call for instance $instanceId was not handled.");
-        }
-
-        // finish the remote call
-        if ($result) {
-            $remoteCallHandler->exitSuccess();
-        } else {
-            $remoteCallHandler->exitFailure();
-        }
-
-        $this->exitRemoteCall();
-    }
-
-    /**
-     * Exit the script
-     *
-     * if a remote call was not handled by the user we die hard here
-     */
-    protected function exitRemoteCall(): void
-    {
-        error_log("Warning: The exit procedure for a remote call was not properly handled.");
-        exit;
-    }
-
     /**
      * Strips the string "Bearer " from the Authorization header value and returns the JWT token
      * @param string $value
@@ -206,11 +166,11 @@ class HeaderSession
         return substr($value, 7);
     }
 
-	/**
-	 * Checks if the request is made as an editor
-	 *
-	 * @return bool
-	 */
+    /**
+     * Checks if the request is made as an editor
+     *
+     * @return bool
+     */
     private function isAdminView(): bool
     {
         return $this->isEditor() && URLHelper::getParam(self::EDITOR_STUDIO_PATH);
